backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
motor==3.3.2
pydantic==2.6.1
python-multipart==0.0.9
google-generativeai==0.3.2
python-dotenv==1.0.1
websockets==12.0
python-jose==3.3.0 